reverted unneeded dependency introduction
